+++ src/boxes/media_inforamtion.rs
@@ -5,7 +5,7 @@ use super::sample_table::SampleTableBox;
 
 const MEDIA_INFORMATION_BOX_HEADER_SIZE: usize = 8; // Size of BoxHeader (fixed)
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct MediaInformationBox {
     header: BoxHeader, // Size and type at offset 0–7
     stbl: SampleTableBox,
@@ -39,7 +39,7 @@ impl ReadHelper for MediaInformationBox {
 
     /// Calculates the total size of the MediaInformationBox in bytes, including the BoxHeader and SampleTableBox fields.
     fn total_size(&self) -> usize {
-        let header_size = self.header.total_size() as usize;
+        let header_size = self.header.total_size();
         let stbl_size = self.stbl.total_size(); // Size of SampleTableBox
 
         header_size + stbl_size // Total size is the sum of both components
