+++ src/boxes/time_to_sample.rs
@@ -9,7 +9,7 @@ const TIME_TO_SAMPLE_BOX_ENTRY_COUNT_SIZE: usize = 4; // 4 bytes for entry_count
 const TIME_TO_SAMPLE_BOX_HEADER_SIZE: usize = 8; // Size of BoxHeader (fixed part)
 const TIME_TO_SAMPLE_BOX_ENTRY_SIZE: usize = 8; // 8 bytes for each entry (sample_count + duration)
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct TimeToSampleBox {
     header: BoxHeader,        // Size and type at offset 0–7
     entry_count: u32,         // 4 bytes at offset 8–11
@@ -87,7 +87,7 @@ impl ReadHelper for TimeToSampleBox {
     }
 
     fn total_size(&self) -> usize {
-        let header_size = self.header.total_size() as usize; // Size of the BoxHeader
+        let header_size = self.header.total_size(); // Size of the BoxHeader
         let entry_count_size = TIME_TO_SAMPLE_BOX_ENTRY_COUNT_SIZE; // Size of entry_count (4 bytes)
         let entry_size = self.entries.len() * TIME_TO_SAMPLE_BOX_ENTRY_SIZE; // Variable size based on entries
